miner actor state
