+++ core/vm/actor/builtin/states/CMakeLists.txt
@@ -20,3 +20,4 @@ add_subdirectory(account)
 add_subdirectory(cron)
 add_subdirectory(init)
 add_subdirectory(market)
+add_subdirectory(miner)
