+++ core/vm/actor/builtin/states/miner/v2/miner_actor_state.hpp
@@ -5,12 +5,12 @@
 
 #pragma once
 
-#include "vm/actor/builtin/states/miner_actor_state.hpp"
+#include "vm/actor/builtin/states/miner/miner_actor_state.hpp"
 
 #include "codec/cbor/streams_annotation.hpp"
 #include "storage/ipfs/datastore.hpp"
 
-namespace fc::vm::actor::builtin::v3::miner {
+namespace fc::vm::actor::builtin::v2::miner {
   using types::Universal;
 
   struct MinerActorState : states::MinerActorState {
@@ -38,13 +38,13 @@ namespace fc::vm::actor::builtin::v3::miner {
              deadlines,
              early_terminations)
 
-}  // namespace fc::vm::actor::builtin::v3::miner
+}  // namespace fc::vm::actor::builtin::v2::miner
 
 namespace fc::cbor_blake {
   template <>
-  struct CbVisitT<vm::actor::builtin::v3::miner::MinerActorState> {
+  struct CbVisitT<vm::actor::builtin::v2::miner::MinerActorState> {
     template <typename Visitor>
-    static void call(vm::actor::builtin::v3::miner::MinerActorState &state,
+    static void call(vm::actor::builtin::v2::miner::MinerActorState &state,
                      const Visitor &visit) {
       visit(state.miner_info);
       visit(state.vesting_funds);
