+++ core/vm/actor/builtin/v0/miner/CMakeLists.txt
@@ -5,10 +5,10 @@
 
 add_library(miner_actor_v0
     miner_actor.cpp
-    miner_actor_state.cpp
     miner_actor_utils.cpp
     )
 target_link_libraries(miner_actor_v0
+    miner_actor_state
     actor
     address
     array
