+++ core/vm/actor/builtin/v3/miner/CMakeLists.txt
@@ -5,10 +5,9 @@
 
 add_library(miner_actor_v3
     miner_actor.cpp
-    miner_actor_state.cpp
     miner_actor_utils.cpp
     )
 target_link_libraries(miner_actor_v3
+    miner_actor_state
     miner_actor_v2
-    miner_types
     )
