+++ core/vm/actor/builtin/v3/miner/miner_actor.cpp
@@ -5,7 +5,7 @@
 
 #include "vm/actor/builtin/v3/miner/miner_actor.hpp"
 
-#include "vm/actor/builtin/states/miner_actor_state.hpp"
+#include "vm/actor/builtin/states/miner/miner_actor_state.hpp"
 #include "vm/actor/builtin/types/type_manager/type_manager.hpp"
 #include "vm/actor/builtin/v3/account/account_actor.hpp"
 #include "vm/actor/builtin/v3/storage_power/storage_power_actor_export.hpp"
